requirements.txt
certifi==2019.6.16
chardet==3.0.4
click==6.6
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==0.23
pyaml==16.9.0
PyYAML==5.1.2
requests==2.22.0
six==1.10.0
slackclient==2.1.0
urllib3==1.25.3
websocket-client==0.37.0
Werkzeug==0.15.5
